+++ src/routes/navigation/navigation.styles.jsx
@@ -5,7 +5,7 @@
  * Author: Paul Adrian Reyes (paulreyes74@yahoo.com)
  * GitHub: https://github.com/alcoranpaul
  * -----
- * Last Modified: Thursday, 29th June 2023 2:24:04 pm
+ * Last Modified: Thursday, 29th June 2023 2:34:52 pm
  * Modified By: PR (paulreyes74@yahoo.com>)
  * -----
  * -----
@@ -71,18 +71,3 @@ export const NavItem = styled(Link)`
 `
 
 
-export const AdminSignIn = styled(Link)`
-    position: absolute; 
-    right: 40px;
-    bottom: 50px;
-    background-color: var(--color-background);
-    border: 2px double var(--color-accent);
-    color: var(--color-text);
-    opacity: 0.4;
-    padding: 10px;
-    text-decoration: none;
-    transition: all 0.2s ease-in-out;
-    &:hover {
-      opacity: 1;
-    }
-`
